Handle broken aiohttp websocket in chat bot client
It turns out that aiohttp WebSocket code does not recognize if the
underlying connection is broken:
https://github.com/aio-libs/aiohttp/issues/2309
As a twitch client we know we'll at minimum receive a PING request every
5 minutes even if no other channel updates are happening:
https://dev.twitch.tv/docs/irc/#keepalive-messages
(I tested this and saw every a ping request always happened between
every 4min 1sec to 4min 51sec)

If we miss a PING and fail to respond twitch considers the connection
closed. Thus we can add a timeout to our websocket receive() call and if
we've heard nothing in over 5 minutes our connection is broken and we
should perform a reconnect. This has been tested running the bot, and
then trying 3 different types of connection breaking:
1) Disabling the network interface on my computer for 5 minutes
2) Suspending laptop and resuming later
3) Dropping related active NAT state on router (similar to rebooting router)