Rewrite session_reconnect handling to follow twitch rules
https://dev.twitch.tv/docs/eventsub/handling-websocket-events/#reconnect-message
When a session_reconnect message is received, the client should create
a new connection to the reconnect url while leaving the existing
connection open to receive messages until the new connection gets
the welcome message. This way no events should be missed during the
handoff.

In my testing this fixes #286 where the connection gets broken